requirements.txt
# GründerAI Backend Dependencies - MVP Version

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# CORS and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis for session management
redis==5.0.1
hiredis==2.2.3

# Data processing (for later IRT integration)
numpy==1.26.2
pandas==2.1.3
scipy==1.11.4

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Logging and monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2